add login restrictions

+++ catalog.py
@@ -82,8 +82,8 @@ def createUser():
 # log out
 @app.route('/logout')
 def logOut():
-    del login_session['username']  
     session.close()
+    del login_session['username']  
     return redirect(url_for('getMainPage'))
 
 
@@ -96,32 +96,29 @@ def getMainPage():
     latestParts = session.query(Parts).order_by(Parts.id.desc()).limit(10)
     categories = session.query(Categories).group_by(Categories.name)
     users = session.query(Users).all()
-    if 'username' in login_session:
-        user = login_session['username']
-        session.close()
-        return render_template('mainPrivate.html', categories = categories, latestParts = latestParts, user = user )
-    return render_template('main.html', categories = categories, latestParts = latestParts, users = users)
+    if 'username' not in login_session:
+        return render_template('main.html', categories = categories, latestParts = latestParts, users = users)
+    return render_template('mainPrivate.html', categories = categories, latestParts = latestParts, user = login_session['username'] )
+    
 
 # new part page --shows form to set up a new part
 @app.route('/newpart', methods=['GET', 'POST'])
 def createNewPart():
     session.close()
     categories = session.query(Categories).group_by(Categories.name)
-    if 'username' in login_session:
-        user = login_session['username']
-    else:
+    if 'username' not in login_session:
         return redirect(url_for('showLogin'))
     if request.method == 'POST':
         partName = request.form['partName']
         partDescription = request.form['partDescription']
         partCategoryName = request.form['partCategory']
         partCategory = session.query(Categories).filter_by(name=partCategoryName).first()
-        partCreator = session.query(Users).first()
+        partCreator = session.query(Users).filter_by(username = user).first()
         partToAdd = Parts(name=partName, description = partDescription, category_id = partCategory.id, user_id = partCreator.id)
         session.add(partToAdd)
         session.commit()
         return redirect(url_for('getMainPage'))
-    return render_template('newPart.html', categories = categories, user = user)
+    return render_template('newPart.html', categories = categories, user = login_session['username'] )
 
 # category page --  shows all parts of the specific category
 @app.route('/<int:category_id>')
@@ -130,20 +127,30 @@ def getCategory(category_id):
     categories = session.query(Categories).group_by(Categories.name)
     requestedCategory = session.query(Categories).filter_by(id=category_id).first()
     partsOfrequestedCategory = session.query(Parts).filter_by(category_id = requestedCategory.id)
-    return render_template('category.html', categories = categories, category = requestedCategory, parts=partsOfrequestedCategory)
+    if 'username' not in login_session:
+        return render_template('category.html',categories = categories, category = requestedCategory, parts=partsOfrequestedCategory)
+    return render_template('categoryPrivate.html',categories = categories, category = requestedCategory, parts=partsOfrequestedCategory,user = login_session['username'])
 
 # part page --shows all info of the specific part
 @app.route('/<int:category_id>/<int:part_id>')
 def getPart(category_id, part_id):
     session.close()
-    requestedPart = session.query(Parts).filter_by(id=part_id).first()    
+    requestedPart = session.query(Parts).filter_by(id=part_id).first()  
+    if 'username' in login_session:
+        user = login_session['username']
+        return render_template('partPrivate.html', part = requestedPart, user = user)  
     return render_template('part.html', part = requestedPart)
 
 # part edit page --shows form to edit a part
 @app.route('/<int:category_id>/<int:part_id>/edit', methods=['GET', 'POST'] )
 def editPart(category_id, part_id):
     session.close()
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     partToedit = session.query(Parts).filter_by(id=part_id).first()
+    user = session.query(Users).filter_by(id=partToedit.user_id).first()
+    if user.username != login_session.get('username'):
+        return redirect(url_for('getPart', category_id = category_id, part_id = part_id))
     if (request.method == "POST"):
         partName = request.form['partName']
         partDescription = request.form['partDescription']
@@ -153,20 +160,25 @@ def editPart(category_id, part_id):
         categoryOfPart = session.query(Categories).filter_by(name=partCategory).first()
         partToedit.category_id = categoryOfPart.id
         session.commit()
-        return redirect(url_for('getPart',category_id = category_id, part_id = part_id ))
+        return redirect(url_for('getPart',category_id = partToedit.category_id, part_id = part_id ))
     categories = session.query(Categories).group_by(Categories.name)
-    return render_template('editPart.html', categories = categories, part = partToedit)
+    return render_template('editPart.html', categories = categories, part = partToedit, user = user.username)
 
 # part delete page --shows form to delete a part
 @app.route('/<int:category_id>/<int:part_id>/delete', methods=['GET','POST'] )
 def deletePart(category_id,part_id):
     session.close()
+    if 'username' not in login_session:
+        return redirect(url_for('showLogin'))
     partToDelete = session.query(Parts).filter_by(id=part_id).first()
+    user = session.query(Users).filter_by(id=partToDelete.user_id).first()
+    if user.username != login_session.get('username'):
+        return redirect(url_for('getPart', category_id = category_id, part_id = part_id))
     if request.method == 'POST':
         session.delete(partToDelete)
         session.commit()
         return redirect(url_for('getCategory', category_id = category_id))
-    return render_template('deletePart.html', part = partToDelete)
+    return render_template('deletePart.html', part = partToDelete, user = login_session['username'])
 
 
 if (__name__ == '__main__'):
